+++ engine/src/androidTest/java/com/google/android/fhir/db/impl/DatabaseImplTest.kt
@@ -18,6 +18,7 @@ package com.google.android.fhir.db.impl
 
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import ca.uhn.fhir.rest.param.ParamPrefixEnum
 import com.google.android.fhir.FhirServices
 import com.google.android.fhir.db.ResourceNotFoundInDbException
 import com.google.android.fhir.db.impl.entities.LocalChangeEntity
@@ -28,10 +29,12 @@ import com.google.android.fhir.search.Search
 import com.google.android.fhir.search.StringFilterModifier
 import com.google.android.fhir.search.getQuery
 import com.google.android.fhir.sync.FhirDataSource
+import com.google.common.truth.Truth
 import com.google.common.truth.Truth.assertThat
 import java.math.BigDecimal
 import kotlinx.coroutines.runBlocking
 import org.hl7.fhir.r4.model.Bundle
+import org.hl7.fhir.r4.model.DateTimeType
 import org.hl7.fhir.r4.model.DecimalType
 import org.hl7.fhir.r4.model.Enumerations
 import org.hl7.fhir.r4.model.HumanName
@@ -503,6 +506,344 @@ class DatabaseImplTest {
     }
 
     assertThat(res).hasSize(0)
+  }
+
+  @Test
+  fun search_date_starts_after() {
+    val patient1 =
+      Patient().apply {
+        id = "1"
+        deceased = DateTimeType("2013-03-25")
+      }
+    val patient2 =
+      Patient().apply {
+        id = "2"
+        deceased = DateTimeType("2013-03-10")
+      }
+    val patient3 =
+      Patient().apply {
+        id = "3"
+        deceased = DateTimeType("2013-03-03")
+      }
+    val patient4 =
+      Patient().apply {
+        id = "4"
+        deceased = DateTimeType("2013-03-23T10:00:00")
+      }
+    val res = runBlocking {
+      database.insert(patient1, patient2, patient3, patient4)
+      database.search<Patient>(
+        Search(ResourceType.Patient)
+          .apply {
+            filter(Patient.DEATH_DATE) {
+              prefix = ParamPrefixEnum.STARTS_AFTER
+              value = DateTimeType("2013-03-14")
+            }
+          }
+          .getQuery()
+      )
+    }
+    Truth.assertThat(res).hasSize(2)
+    Truth.assertThat(
+        res.all { it.deceasedDateTimeType.value.time >= DateTimeType("2013-03-14").value.time }
+      )
+      .isTrue()
+  }
+  @Test
+  fun search_date_ends_before() {
+    val patient1 =
+      Patient().apply {
+        id = "1"
+        deceased = DateTimeType("2013-03-25")
+      }
+    val patient2 =
+      Patient().apply {
+        id = "2"
+        deceased = DateTimeType("2013-03-10")
+      }
+    val patient3 =
+      Patient().apply {
+        id = "3"
+        deceased = DateTimeType("2013-03-03")
+      }
+    val patient4 =
+      Patient().apply {
+        id = "4"
+        deceased = DateTimeType("2013-03-23T10:00:00")
+      }
+    val res = runBlocking {
+      database.insert(patient1, patient2, patient3, patient4)
+      database.search<Patient>(
+        Search(ResourceType.Patient)
+          .apply {
+            filter(Patient.DEATH_DATE) {
+              prefix = ParamPrefixEnum.ENDS_BEFORE
+              value = DateTimeType("2013-03-14")
+            }
+          }
+          .getQuery()
+      )
+    }
+    Truth.assertThat(res).hasSize(2)
+    Truth.assertThat(
+        res.all { it.deceasedDateTimeType.value.time <= DateTimeType("2013-03-14").value.time }
+      )
+      .isTrue()
+  }
+
+  @Test
+  fun search_date_not_equals() {
+    val patient1 =
+      Patient().apply {
+        id = "1"
+        deceased = DateTimeType("2013-03-25")
+      }
+    val patient2 =
+      Patient().apply {
+        id = "2"
+        deceased = DateTimeType("2013-03-14")
+      }
+    val patient3 =
+      Patient().apply {
+        id = "3"
+        deceased = DateTimeType("2013-03-15")
+      }
+    val patient4 =
+      Patient().apply {
+        id = "4"
+        deceased = DateTimeType("2013-03-14T23:59:59.999")
+      }
+    val res = runBlocking {
+      database.insert(patient1, patient2, patient3, patient4)
+      database.search<Patient>(
+        Search(ResourceType.Patient)
+          .apply {
+            filter(Patient.DEATH_DATE) {
+              prefix = ParamPrefixEnum.NOT_EQUAL
+              value = DateTimeType("2013-03-14")
+            }
+          }
+          .getQuery()
+      )
+    }
+    Truth.assertThat(res).hasSize(2)
+    Truth.assertThat(
+        res.all {
+          it.deceasedDateTimeType.value.time >= DateTimeType("2013-03-15").value.time ||
+            it.deceasedDateTimeType.value.time < DateTimeType("2013-03-14").value.time
+        }
+      )
+      .isTrue()
+  }
+
+  @Test
+  fun search_date_equals() {
+    val patient1 =
+      Patient().apply {
+        id = "1"
+        deceased = DateTimeType("2013-03-25")
+      }
+    val patient2 =
+      Patient().apply {
+        id = "2"
+        deceased = DateTimeType("2013-03-14")
+      }
+    val patient3 =
+      Patient().apply {
+        id = "3"
+        deceased = DateTimeType("2013-03-15")
+      }
+    val patient4 =
+      Patient().apply {
+        id = "4"
+        deceased = DateTimeType("2013-03-14T23:59:59.999")
+      }
+    val res = runBlocking {
+      database.insert(patient1, patient2, patient3, patient4)
+      database.search<Patient>(
+        Search(ResourceType.Patient)
+          .apply {
+            filter(Patient.DEATH_DATE) {
+              prefix = ParamPrefixEnum.EQUAL
+              value = DateTimeType("2013-03-14")
+            }
+          }
+          .getQuery()
+      )
+    }
+    Truth.assertThat(res).hasSize(2)
+    Truth.assertThat(
+        res.all {
+          it.deceasedDateTimeType.value.time < DateTimeType("2013-03-15").value.time &&
+            it.deceasedDateTimeType.value.time < DateTimeType("2013-03-15").value.time
+        }
+      )
+      .isTrue()
+  }
+  @Test
+  fun search_date_greater() {
+    val patient1 =
+      Patient().apply {
+        id = "1"
+        deceased = DateTimeType("2013-03-13")
+      }
+    val patient2 =
+      Patient().apply {
+        id = "2"
+        deceased = DateTimeType("2013-03-14")
+      }
+    val patient3 =
+      Patient().apply {
+        id = "3"
+        deceased = DateTimeType("2013-03-15")
+      }
+    val patient4 =
+      Patient().apply {
+        id = "4"
+        deceased = DateTimeType("2013-03-14T23:59:59.999")
+      }
+    val res = runBlocking {
+      database.insert(patient1, patient2, patient3, patient4)
+      database.search<Patient>(
+        Search(ResourceType.Patient)
+          .apply {
+            filter(Patient.DEATH_DATE) {
+              prefix = ParamPrefixEnum.GREATERTHAN
+              value = DateTimeType("2013-03-14")
+            }
+          }
+          .getQuery()
+      )
+    }
+    Truth.assertThat(res).hasSize(1)
+    Truth.assertThat(
+        res.all { it.deceasedDateTimeType.value.time >= DateTimeType("2013-03-15").value.time }
+      )
+      .isTrue()
+  }
+  @Test
+  fun search_date_greater_or_equal() {
+    val patient1 =
+      Patient().apply {
+        id = "1"
+        deceased = DateTimeType("2013-03-13")
+      }
+    val patient2 =
+      Patient().apply {
+        id = "2"
+        deceased = DateTimeType("2013-03-14")
+      }
+    val patient3 =
+      Patient().apply {
+        id = "3"
+        deceased = DateTimeType("2013-03-15")
+      }
+    val patient4 =
+      Patient().apply {
+        id = "4"
+        deceased = DateTimeType("2013-03-14T23:59:59.999")
+      }
+    val res = runBlocking {
+      database.insert(patient1, patient2, patient3, patient4)
+      database.search<Patient>(
+        Search(ResourceType.Patient)
+          .apply {
+            filter(Patient.DEATH_DATE) {
+              prefix = ParamPrefixEnum.GREATERTHAN_OR_EQUALS
+              value = DateTimeType("2013-03-14")
+            }
+          }
+          .getQuery()
+      )
+    }
+    Truth.assertThat(res).hasSize(3)
+    Truth.assertThat(
+        res.all { it.deceasedDateTimeType.value.time >= DateTimeType("2013-03-14").value.time }
+      )
+      .isTrue()
+  }
+  @Test
+  fun search_date_less() {
+    val patient1 =
+      Patient().apply {
+        id = "1"
+        deceased = DateTimeType("2013-03-13")
+      }
+    val patient2 =
+      Patient().apply {
+        id = "2"
+        deceased = DateTimeType("2013-03-14")
+      }
+    val patient3 =
+      Patient().apply {
+        id = "3"
+        deceased = DateTimeType("2013-03-15")
+      }
+    val patient4 =
+      Patient().apply {
+        id = "4"
+        deceased = DateTimeType("2013-03-14T23:59:59.999")
+      }
+    val res = runBlocking {
+      database.insert(patient1, patient2, patient3, patient4)
+      database.search<Patient>(
+        Search(ResourceType.Patient)
+          .apply {
+            filter(Patient.DEATH_DATE) {
+              prefix = ParamPrefixEnum.LESSTHAN
+              value = DateTimeType("2013-03-14")
+            }
+          }
+          .getQuery()
+      )
+    }
+    Truth.assertThat(res).hasSize(1)
+    Truth.assertThat(
+        res.all { it.deceasedDateTimeType.value.time < DateTimeType("2013-03-14").value.time }
+      )
+      .isTrue()
+  }
+
+  @Test
+  fun search_date_less_or_equal() {
+    val patient1 =
+      Patient().apply {
+        id = "1"
+        deceased = DateTimeType("2013-03-13")
+      }
+    val patient2 =
+      Patient().apply {
+        id = "2"
+        deceased = DateTimeType("2013-03-14")
+      }
+    val patient3 =
+      Patient().apply {
+        id = "3"
+        deceased = DateTimeType("2013-03-15")
+      }
+    val patient4 =
+      Patient().apply {
+        id = "4"
+        deceased = DateTimeType("2013-03-14T23:59:59.999")
+      }
+    val res = runBlocking {
+      database.insert(patient1, patient2, patient3, patient4)
+      database.search<Patient>(
+        Search(ResourceType.Patient)
+          .apply {
+            filter(Patient.DEATH_DATE) {
+              prefix = ParamPrefixEnum.LESSTHAN_OR_EQUALS
+              value = DateTimeType("2013-03-14")
+            }
+          }
+          .getQuery()
+      )
+    }
+    Truth.assertThat(res).hasSize(3)
+    Truth.assertThat(
+        res.all { it.deceasedDateTimeType.value.time <= DateTimeType("2013-03-15").value.time }
+      )
+      .isTrue()
   }
 
   private companion object {
